optional_plugins: Use simple "copy" in "loader_yaml" for args
The use of "deepcopy" should not be necessary as we only use it here to
override certain keys with our values, not to modify their value.

Even if (theoretically) modified those mutable values later in the
code, it'd have been done even without "deepcopy" for cases where we
didn't forced the "args".

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>